requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.1
PyJWT==2.8.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
Werkzeug==3.0.3
whitenoise==6.7.0
